данные и функции для генерации данных перенесены в файл data.js
