+++ js/main.js
@@ -1,16 +1,4 @@
-import {getRandomAvatar} from './random-functions.js';
-import {getRandomOffer} from './random-functions.js';
-import {getRandomObjectLocation} from './random-functions.js';
-
-const IMG_ADDRESS = 'img/avatars/user';
-const getNewBookingObject = () => {
-  return {
-    author: getRandomAvatar(IMG_ADDRESS),
-    offer: getRandomOffer(),
-    objectLocation: getRandomObjectLocation(),
-  };
-};
-
+import {getNewBookingObject} from './data.js';
 
 const OBJECTS_COUNT = 10;
 const bookingObjectsList = new Array(OBJECTS_COUNT).fill(null).map(() => getNewBookingObject());
